deadend
stop/start takes a lot of time + after stop plays from the beginning, but not from current stream moment.
pause/start works fine but not from beginning
both have problem that current song name is not the name of the song playing because of huge delay which is got when pause
That's why will try ExoPlayer instead of MediaPlayer.